Refactor Google Drive file retrieval
